Allow for 204 responses from Sirius
The new Access code letter request endpoint returns  a 204 with no body content. When using the `requests` library an exception is thrown when calling `.json()` on a response that has the 204 code - even though the response was valid the library considers it to be in error. This PR catches the 204 case and solves the issue.